print doc updated

+++ wp-auto-install.sh
@@ -9,6 +9,16 @@
 #5. extract and copy wordpress to /var/www/html/sitename.com
 #6. create database and create user with given details in step 2
 #7. modify nginx file and restart nginx
+print_version_info(){
+	cat <<END
+	1. Php version : 		$1
+	2. Wordpress version: 	$2
+	3. Database name: 		$3
+	4. Datbase usernmae: 	$4
+	4. Database password: 	$5
+	5. Site Domain Name: 	$6
+END
+}
 run_sql_commands(){
 
 	dbname=$1
@@ -29,29 +39,50 @@ run_sql_commands(){
 	
 }
 
+# 
+# script start
+# 
+
+echo -e "\n\n1.) Which php version you want to intall?"
+read vphp
+
+echo -e "\n2.) Which wordpress version you want to install?"
+read vwp
+
+echo -e "\n3.) Please enter your database name."
+read dbname
+
+echo -e "\n4.) Please enter database username."
+read dbuname
+
+echo -e "\n5.) Please enter database password."
+read dbpwd
+
+echo -e "\n6.) Please enter your site domain name. Example: geekconsults.co.in "
+read sitename
+
+echo -e "\n\nPlease keep the above entered information handy and keep it handy as it will required during wordpress setup."
+
+print_version_info $vphp $vwp $dbname $dbuname $dbpwd $sitename
+
+echo -e "\n\nAre you sure you want to continue installation? y/n"
+read -n 1 confirm_install
+
+#exit if not confirm installation
+if [[ "$confirm_install" != "y" ]]
+	echo "\n\nProgram Terminated..."
+	exit 0
+fi
+
+# 
+#installtion of dependencies
+# 
+
 # php 7.1 wp 5.2.3 5.3
 sudo apt-get -y install software-properties-common
 sudo add-apt-repository ppa:ondrej/php
 sudo apt -y update
 sudo apt -y upgrade
-
-echo "php version?"
-read vphp
-
-echo "wordpress version ?"
-read vwp
-
-echo "database name?"
-read dbname
-
-echo "database username?"
-read dbuname
-
-echo "database password?"
-read dbpwd
-
-echo "Site name?"
-read sitename
 
 echo "Installing php${vphp}.."
 sudo apt -y install php${vphp}-cli php${vphp}-fpm php${vphp}-mysql php${vphp}-json php${vphp}-opcache php${vphp}-mbstring php${vphp}-xml php${vphp}-gd php${vphp}-curl
@@ -87,7 +118,7 @@ sudo chown -R www-data: /var/www/html/$sitename
 cd $default_dir
 cp nginx.conf /tmp/$sitename
 sed -i "s/domain.tld/$sitename/g" /tmp/$sitename
-sed -i "s/vphp/$vphp/g" /tmp/$sitename    
+sed -i "s/_vphp_/$vphp/g" /tmp/$sitename    
 sudo mv /tmp/$sitename /etc/nginx/sites-available/                                                      
 # cd /etc/nginx/sites-available
 sudo ln -s /etc/nginx/sites-available/$sitename /etc/nginx/sites-enabled/
@@ -95,7 +126,10 @@ sudo ln -s /etc/nginx/sites-available/$sitename /etc/nginx/sites-enabled/
 echo "Hurray!! All done. Restarting server..."
 sudo systemctl restart nginx
 
-echo -e "\nWordpress setup completed successfully!"
+echo -e "\nWordpress installation completed successfully!"
+echo -e "\n\nPlease keep the below information safe for future reference."
+print_version_info $vphp $vwp $dbname $dbuname $dbpwd $sitename
+
 
 #copy all creds and version in a text file and tell about that to user
